фикс

+++ internal/domain/models/withdrawals.go
@@ -6,5 +6,5 @@ type Withdrawal struct {
 	ProcessedAt         time.Time `json:"-"`
 	ProcessedAtFormated string    `json:"processed_at"`
 	OrderNumber         string    `json:"order"`
-	Sum                 int64     `json:"sum"`
+	Sum                 float64   `json:"sum"`
 }
